Rearranging code
Change db.go package from main to dbs, shifting it to a directory of dbs.

Create service.go which is now the main package.  Imports dbs and uses methods in db.go.

Renamed methods in db.go, so they become available to other packages.  Seems to export them, methods need to star with a uppercase letter.